ISSUE-782 Move JmapSettingsProbe to dedicated probe package
